refactor: split napi lint helpers
