+++ crates/ox_content_napi/src/lint.rs
@@ -9,6 +9,20 @@ use regex::Regex;
 use serde::Deserialize;
 use unicode_normalization::UnicodeNormalization;
 
+mod cjk;
+mod diagnostics;
+mod dictionary;
+mod latin;
+mod mask;
+mod tokens;
+mod utils;
+
+use diagnostics::*;
+use dictionary::*;
+use mask::*;
+use tokens::*;
+use utils::*;
+
 const SUPPORTED_MARKDOWN_LINT_LANGUAGES: [&str; 6] = ["en", "ja", "zh", "fr", "de", "pl"];
 const DEFAULT_LANGUAGES: [&str; 1] = ["en"];
 
@@ -526,922 +540,4 @@ fn collect_markdown_lint_state(
     }
 
     MarkdownLintState { diagnostics, masked_lines }
-}
-
-fn create_dictionary_bundle(options: &InternalMarkdownLintOptions) -> DictionaryBundle {
-    let extra_global_words = options
-        .dictionary
-        .words
-        .iter()
-        .map(|word| normalize_word_for_set(word))
-        .filter(|word| !word.is_empty())
-        .collect::<FxHashSet<_>>();
-
-    let extra_by_language = options
-        .dictionary
-        .by_language
-        .iter()
-        .map(|(language, words)| {
-            (
-                language.clone(),
-                words
-                    .iter()
-                    .map(|word| normalize_word_for_set(word))
-                    .filter(|word| !word.is_empty())
-                    .collect::<FxHashSet<_>>(),
-            )
-        })
-        .filter(|(_, words)| !words.is_empty())
-        .collect::<FxHashMap<_, _>>();
-
-    let mut latin_words = FxHashSet::default();
-    let mut latin_suggestion_words = FxHashSet::default();
-
-    for language in &options.languages {
-        if language == "ja" || language == "zh" {
-            continue;
-        }
-
-        if let Some(words) = PREPARED_LINT_DICTIONARY_DATA.by_language.get(language.as_str()) {
-            for word in &words.words {
-                latin_words.insert(word.clone());
-                latin_suggestion_words.insert(word.clone());
-            }
-        }
-
-        for word in &PREPARED_LINT_DICTIONARY_DATA.global_words {
-            latin_words.insert(word.clone());
-            latin_suggestion_words.insert(word.clone());
-        }
-
-        if let Some(words) = extra_by_language.get(language.as_str()) {
-            for word in words {
-                latin_words.insert(word.clone());
-                latin_suggestion_words.insert(word.clone());
-            }
-        }
-    }
-
-    for word in &extra_global_words {
-        latin_words.insert(word.clone());
-        latin_suggestion_words.insert(word.clone());
-    }
-
-    let ignored_words =
-        options.dictionary.ignored_words.iter().map(|word| normalize_word_for_set(word)).collect();
-
-    let active_languages = options
-        .languages
-        .iter()
-        .filter(|language| {
-            PREPARED_LINT_DICTIONARY_DATA
-                .by_language
-                .get((*language).as_str())
-                .is_some_and(|entry| entry.has_base_words)
-                || extra_by_language
-                    .get((*language).as_str())
-                    .is_some_and(|words| !words.is_empty())
-        })
-        .cloned()
-        .collect();
-
-    let mut cjk_segment_words = FxHashMap::default();
-
-    for language in
-        options.languages.iter().filter(|language| *language == "ja" || *language == "zh")
-    {
-        let mut words = PREPARED_LINT_DICTIONARY_DATA
-            .by_language
-            .get(language.as_str())
-            .map_or_else(Vec::new, |entry| entry.cjk_segment_words.clone());
-
-        words.extend(
-            extra_global_words
-                .iter()
-                .filter(|word| word.chars().any(is_cjk_char))
-                .map(|word| SegmentWord { char_len: count_code_points(word), text: word.clone() }),
-        );
-
-        if let Some(extra_words) = extra_by_language.get(language.as_str()) {
-            words.extend(
-                extra_words.iter().filter(|word| word.chars().any(is_cjk_char)).map(|word| {
-                    SegmentWord { char_len: count_code_points(word), text: word.clone() }
-                }),
-            );
-        }
-
-        sort_and_dedupe_segment_words(&mut words);
-        if !words.is_empty() {
-            cjk_segment_words.insert(language.clone(), words);
-        }
-    }
-
-    DictionaryBundle {
-        active_languages,
-        cjk_segment_words,
-        extra_by_language,
-        extra_global_words,
-        ignored_words,
-        latin_words,
-        latin_suggestion_words: {
-            let mut values = latin_suggestion_words.into_iter().collect::<Vec<_>>();
-            values.sort();
-            values
-        },
-    }
-}
-
-fn collect_tokens(
-    masked_line: &str,
-    languages: &[String],
-    dictionary: &DictionaryBundle,
-) -> Vec<Token> {
-    let latin_tokens = collect_latin_tokens(masked_line, languages, dictionary);
-    let mut cjk_tokens = Vec::new();
-
-    if let Some(cjk_run_pattern) = CJK_RUN_PATTERN.as_ref() {
-        for value in cjk_run_pattern.find_iter(masked_line) {
-            let start = byte_to_char_index(masked_line, value.start());
-            cjk_tokens.extend(collect_cjk_tokens(value.as_str(), start, languages, dictionary));
-        }
-    }
-
-    merge_tokens(latin_tokens, cjk_tokens)
-}
-
-fn merge_tokens(left: Vec<Token>, right: Vec<Token>) -> Vec<Token> {
-    let mut merged = Vec::with_capacity(left.len() + right.len());
-    let mut left_iter = left.into_iter().peekable();
-    let mut right_iter = right.into_iter().peekable();
-
-    while left_iter.peek().is_some() || right_iter.peek().is_some() {
-        let take_left = match (left_iter.peek(), right_iter.peek()) {
-            (Some(left_token), Some(right_token)) => left_token.start <= right_token.start,
-            (Some(_), None) => true,
-            _ => false,
-        };
-
-        if take_left {
-            if let Some(token) = left_iter.next() {
-                merged.push(token);
-            }
-        } else if let Some(token) = right_iter.next() {
-            merged.push(token);
-        }
-    }
-
-    merged
-}
-
-fn collect_latin_tokens(
-    masked_line: &str,
-    languages: &[String],
-    dictionary: &DictionaryBundle,
-) -> Vec<Token> {
-    let latin_languages = languages
-        .iter()
-        .filter(|language| language.as_str() != "ja" && language.as_str() != "zh")
-        .cloned()
-        .collect::<Vec<_>>();
-
-    if latin_languages.is_empty() {
-        return Vec::new();
-    }
-
-    let fallback_language = latin_languages[0].clone();
-    let mut tokens = Vec::new();
-
-    if let Some(latin_word_pattern) = LATIN_WORD_PATTERN.as_ref() {
-        for value in latin_word_pattern.find_iter(masked_line) {
-            let text = value.as_str().to_string();
-            let start = byte_to_char_index(masked_line, value.start());
-            let end = start + count_code_points(value.as_str());
-            tokens.push(Token { end, language: fallback_language.clone(), start, text });
-        }
-    }
-
-    assign_latin_languages(tokens, &latin_languages, dictionary, &fallback_language)
-}
-
-fn collect_cjk_tokens(
-    run: &str,
-    start_offset: usize,
-    languages: &[String],
-    dictionary: &DictionaryBundle,
-) -> Vec<Token> {
-    let has_kana =
-        run.chars().any(|value| matches!(value, '\u{3040}'..='\u{309F}' | '\u{30A0}'..='\u{30FF}'));
-
-    let mut candidates = Vec::new();
-
-    if has_kana && languages.iter().any(|language| language == "ja") {
-        candidates.push("ja".to_string());
-    } else {
-        if languages.iter().any(|language| language == "zh") {
-            candidates.push("zh".to_string());
-        }
-        if languages.iter().any(|language| language == "ja") {
-            candidates.push("ja".to_string());
-        }
-    }
-
-    if candidates.is_empty() {
-        return Vec::new();
-    }
-
-    let mut best_candidate: Option<(usize, Vec<Token>)> = None;
-
-    for language in candidates {
-        let tokens = segment_cjk_run(run, start_offset, &language, dictionary);
-        let known_count = tokens.iter().filter(|token| is_known_token(token, dictionary)).count();
-
-        match &best_candidate {
-            Some((best_known_count, best_tokens))
-                if known_count < *best_known_count
-                    || (known_count == *best_known_count && tokens.len() >= best_tokens.len()) => {}
-            _ => best_candidate = Some((known_count, tokens)),
-        }
-    }
-
-    best_candidate.map_or_else(Vec::new, |(_, tokens)| tokens)
-}
-
-fn segment_cjk_run(
-    run: &str,
-    start_offset: usize,
-    language: &str,
-    dictionary: &DictionaryBundle,
-) -> Vec<Token> {
-    let Some(words) = dictionary.cjk_segment_words.get(language) else {
-        return vec![Token {
-            end: start_offset + count_code_points(run),
-            language: language.to_string(),
-            start: start_offset,
-            text: run.to_string(),
-        }];
-    };
-
-    let char_boundaries = collect_char_boundaries(run);
-    let total_chars = char_boundaries.len().saturating_sub(1);
-    let mut tokens = Vec::new();
-    let mut char_index = 0;
-
-    while char_index < total_chars {
-        let start_byte = char_boundaries[char_index];
-        let best_match = words.iter().find(|word| {
-            let end_char = char_index + word.char_len;
-            if end_char > total_chars {
-                return false;
-            }
-
-            let end_byte = char_boundaries[end_char];
-            run[start_byte..end_byte] == word.text
-        });
-
-        if let Some(word) = best_match {
-            tokens.push(Token {
-                end: start_offset + char_index + word.char_len,
-                language: language.to_string(),
-                start: start_offset + char_index,
-                text: word.text.clone(),
-            });
-            char_index += word.char_len;
-            continue;
-        }
-
-        let end_char = char_index + 1;
-        let end_byte = char_boundaries[end_char];
-        tokens.push(Token {
-            end: start_offset + end_char,
-            language: language.to_string(),
-            start: start_offset + char_index,
-            text: run[start_byte..end_byte].to_string(),
-        });
-        char_index = end_char;
-    }
-
-    if tokens.iter().all(|token| count_code_points(&token.text) == 1) {
-        return vec![Token {
-            end: start_offset + count_code_points(run),
-            language: language.to_string(),
-            start: start_offset,
-            text: run.to_string(),
-        }];
-    }
-
-    tokens
-}
-
-fn assign_latin_languages(
-    tokens: Vec<Token>,
-    languages: &[String],
-    dictionary: &DictionaryBundle,
-    fallback_language: &str,
-) -> Vec<Token> {
-    let mut scores =
-        languages.iter().map(|language| (language.clone(), 0_usize)).collect::<FxHashMap<_, _>>();
-
-    let matching_languages = tokens
-        .iter()
-        .map(|token| get_matching_latin_languages(&token.text, languages, dictionary))
-        .collect::<Vec<_>>();
-
-    for matches in &matching_languages {
-        if matches.len() == 1 {
-            let language = &matches[0];
-            *scores.entry(language.clone()).or_default() += 1;
-        }
-    }
-
-    let dominant_language = scores
-        .into_iter()
-        .max_by(|left, right| left.1.cmp(&right.1))
-        .map_or_else(|| fallback_language.to_string(), |(language, _)| language);
-
-    tokens
-        .into_iter()
-        .zip(matching_languages)
-        .map(|(token, matching_languages)| {
-            let inferred_language = matching_languages
-                .first()
-                .cloned()
-                .or_else(|| infer_latin_language_from_characters(&token.text, languages));
-
-            Token {
-                language: inferred_language.unwrap_or_else(|| dominant_language.clone()),
-                ..token
-            }
-        })
-        .collect()
-}
-
-fn get_matching_latin_languages(
-    word: &str,
-    languages: &[String],
-    dictionary: &DictionaryBundle,
-) -> Vec<String> {
-    let normalized_word = normalize_word_for_set(word);
-
-    languages
-        .iter()
-        .filter(|language| {
-            language_contains_word((*language).as_str(), &normalized_word, dictionary)
-        })
-        .cloned()
-        .collect()
-}
-
-fn infer_latin_language_from_characters(word: &str, languages: &[String]) -> Option<String> {
-    if languages.iter().any(|language| language == "pl")
-        && word.chars().any(|value| {
-            matches!(
-                value,
-                'ą' | 'ć'
-                    | 'ę'
-                    | 'ł'
-                    | 'ń'
-                    | 'ó'
-                    | 'ś'
-                    | 'ź'
-                    | 'ż'
-                    | 'Ą'
-                    | 'Ć'
-                    | 'Ę'
-                    | 'Ł'
-                    | 'Ń'
-                    | 'Ó'
-                    | 'Ś'
-                    | 'Ź'
-                    | 'Ż'
-            )
-        })
-    {
-        return Some("pl".to_string());
-    }
-
-    if languages.iter().any(|language| language == "de")
-        && word.chars().any(|value| matches!(value, 'ä' | 'ö' | 'ü' | 'ß' | 'Ä' | 'Ö' | 'Ü'))
-    {
-        return Some("de".to_string());
-    }
-
-    if languages.iter().any(|language| language == "fr")
-        && word.chars().any(|value| {
-            matches!(
-                value,
-                'à' | 'â'
-                    | 'æ'
-                    | 'ç'
-                    | 'é'
-                    | 'è'
-                    | 'ê'
-                    | 'ë'
-                    | 'î'
-                    | 'ï'
-                    | 'ô'
-                    | 'œ'
-                    | 'ù'
-                    | 'û'
-                    | 'ü'
-                    | 'ÿ'
-                    | 'À'
-                    | 'Â'
-                    | 'Æ'
-                    | 'Ç'
-                    | 'É'
-                    | 'È'
-                    | 'Ê'
-                    | 'Ë'
-                    | 'Î'
-                    | 'Ï'
-                    | 'Ô'
-                    | 'Œ'
-                    | 'Ù'
-                    | 'Û'
-                    | 'Ü'
-                    | 'Ÿ'
-            )
-        })
-    {
-        return Some("fr".to_string());
-    }
-
-    None
-}
-
-fn language_contains_word(
-    language: &str,
-    normalized_word: &str,
-    dictionary: &DictionaryBundle,
-) -> bool {
-    dictionary.extra_global_words.contains(normalized_word)
-        || PREPARED_LINT_DICTIONARY_DATA.global_words.contains(normalized_word)
-        || dictionary
-            .extra_by_language
-            .get(language)
-            .is_some_and(|words| words.contains(normalized_word))
-        || PREPARED_LINT_DICTIONARY_DATA
-            .by_language
-            .get(language)
-            .is_some_and(|words| words.words.contains(normalized_word))
-}
-
-fn collect_repeated_punctuation_diagnostics(
-    line_number: usize,
-    masked_line: &str,
-) -> Vec<JsMarkdownLintDiagnostic> {
-    let chars = masked_line.chars().collect::<Vec<_>>();
-    let mut diagnostics = Vec::new();
-    let mut index = 0;
-
-    while index + 1 < chars.len() {
-        let value = chars[index];
-        if !is_repeated_punctuation_char(value) || chars[index + 1] != value {
-            index += 1;
-            continue;
-        }
-
-        let start = index;
-        let mut end = index + 1;
-        while end < chars.len() && chars[end] == value {
-            end += 1;
-        }
-
-        diagnostics.push(create_diagnostic(
-            "repeated-punctuation",
-            format!(
-                "Repeated punctuation \"{}\" looks accidental.",
-                chars[start..end].iter().collect::<String>()
-            ),
-            line_number,
-            start + 1,
-            end + 1,
-            None,
-            None,
-        ));
-        index = end;
-    }
-
-    diagnostics
-}
-
-fn should_ignore_repeated_word_token(token: &Token) -> bool {
-    if token.language == "ja" || token.language == "zh" {
-        return count_code_points(&token.text) <= 1;
-    }
-
-    normalize_comparable_word(&token.text).chars().count() <= 1
-}
-
-fn should_spellcheck_token(token: &Token, dictionary: &DictionaryBundle) -> bool {
-    let normalized = normalize_word_for_set(&token.text);
-
-    if normalized.is_empty() || dictionary.ignored_words.contains(&normalized) {
-        return false;
-    }
-
-    if !dictionary.active_languages.contains(&token.language) {
-        return false;
-    }
-
-    if token.text.contains('_')
-        || token.text.contains('/')
-        || token.text.contains('\\')
-        || token.text.chars().any(char::is_numeric)
-    {
-        return false;
-    }
-
-    if is_uppercase_token(&token.text) {
-        return false;
-    }
-
-    if token.language == "ja" || token.language == "zh" {
-        if token.language == "ja" && token.text.chars().all(is_hiragana) {
-            return count_code_points(&token.text) > 2;
-        }
-
-        return count_code_points(&token.text) > 1;
-    }
-
-    normalize_comparable_word(&token.text).chars().count() > 2
-}
-
-fn is_known_token(token: &Token, dictionary: &DictionaryBundle) -> bool {
-    let normalized = normalize_word_for_set(&token.text);
-    if normalized.is_empty() || dictionary.ignored_words.contains(&normalized) {
-        return true;
-    }
-
-    if token.language == "ja" || token.language == "zh" {
-        return language_contains_word(token.language.as_str(), &normalized, dictionary);
-    }
-
-    dictionary.latin_words.contains(&normalized)
-}
-
-fn suggest_latin_words(word: &str, candidates: &[String]) -> Vec<String> {
-    let normalized_word = normalize_latin_word(word);
-    let first_char = normalized_word.chars().next();
-    let mut suggestions = candidates
-        .iter()
-        .filter(|candidate| {
-            let candidate_length = candidate.chars().count();
-            let word_length = normalized_word.chars().count();
-            candidate_length.abs_diff(word_length) <= 2
-        })
-        .filter(|candidate| candidate.chars().next() == first_char)
-        .map(|candidate| (candidate.clone(), levenshtein(&normalized_word, candidate)))
-        .filter(|(_, distance)| *distance <= 2)
-        .collect::<Vec<_>>();
-
-    suggestions.sort_by(|left, right| left.1.cmp(&right.1).then_with(|| left.0.cmp(&right.0)));
-
-    dedupe_strings(suggestions.into_iter().take(3).map(|(candidate, _)| candidate).collect())
-}
-
-fn summarize_diagnostics(
-    diagnostics: Vec<JsMarkdownLintDiagnostic>,
-    masked_document: String,
-) -> JsMarkdownLintResult {
-    let error_count =
-        diagnostics.iter().filter(|diagnostic| diagnostic.severity == "error").count();
-    let warning_count =
-        diagnostics.iter().filter(|diagnostic| diagnostic.severity == "warning").count();
-    let info_count = diagnostics.iter().filter(|diagnostic| diagnostic.severity == "info").count();
-
-    JsMarkdownLintResult {
-        diagnostics,
-        error_count: error_count as u32,
-        warning_count: warning_count as u32,
-        info_count: info_count as u32,
-        masked_document,
-    }
-}
-
-fn sort_diagnostics(
-    mut diagnostics: Vec<JsMarkdownLintDiagnostic>,
-) -> Vec<JsMarkdownLintDiagnostic> {
-    diagnostics.sort_by(|left, right| {
-        left.line
-            .cmp(&right.line)
-            .then_with(|| left.column.cmp(&right.column))
-            .then_with(|| left.rule_id.cmp(&right.rule_id))
-    });
-    diagnostics
-}
-
-fn create_diagnostic(
-    rule_id: &str,
-    message: String,
-    line: usize,
-    column: usize,
-    end_column: usize,
-    language: Option<String>,
-    suggestions: Option<Vec<String>>,
-) -> JsMarkdownLintDiagnostic {
-    JsMarkdownLintDiagnostic {
-        rule_id: rule_id.to_string(),
-        severity: "warning".to_string(),
-        message,
-        line: line as u32,
-        column: column as u32,
-        end_line: line as u32,
-        end_column: end_column as u32,
-        language,
-        suggestions,
-    }
-}
-
-fn create_skipped_line_mask(line: &str) -> String {
-    " ".repeat(count_code_points(line))
-}
-
-fn get_trailing_whitespace_length(line: &str) -> usize {
-    line.as_bytes().iter().rev().take_while(|value| **value == b' ' || **value == b'\t').count()
-}
-
-fn is_fence_close(line: &str, fence_char: char, fence_length: usize) -> bool {
-    let trimmed = line.trim_start_matches([' ', '\t']);
-    let bytes = trimmed.as_bytes();
-    let fence_byte = fence_char as u8;
-
-    if bytes.len() < fence_length || !bytes[..fence_length].iter().all(|value| *value == fence_byte)
-    {
-        return false;
-    }
-
-    bytes[fence_length..].iter().all(|value| *value == fence_byte)
-}
-
-fn is_indented_code_block_line(line: &str) -> bool {
-    line.starts_with('\t') || line.starts_with("    ")
-}
-
-fn get_visible_text(text: &str) -> String {
-    collapse_whitespace(&mask_markdown_line(text))
-}
-
-fn mask_markdown_line(line: &str) -> String {
-    let line_chars = line.chars().collect::<Vec<_>>();
-    let mut chars = line_chars.clone();
-
-    if let Some(list_prefix_pattern) = LIST_PREFIX_PATTERN.as_ref() {
-        if let Some(prefix_match) = list_prefix_pattern.find(line) {
-            let (start, end) =
-                byte_range_to_char_range(line, prefix_match.start(), prefix_match.end());
-            blank_range(&mut chars, start, end);
-        }
-    }
-
-    if let Some(footnote_pattern) = FOOTNOTE_PATTERN.as_ref() {
-        for value in footnote_pattern.find_iter(line) {
-            let (start, end) = byte_range_to_char_range(line, value.start(), value.end());
-            blank_range(&mut chars, start, end);
-        }
-    }
-
-    if let Some(url_pattern) = URL_PATTERN.as_ref() {
-        for value in url_pattern.find_iter(line) {
-            let (start, end) = byte_range_to_char_range(line, value.start(), value.end());
-            blank_range(&mut chars, start, end);
-        }
-    }
-
-    if let Some(html_tag_pattern) = HTML_TAG_PATTERN.as_ref() {
-        for value in html_tag_pattern.find_iter(line) {
-            let (start, end) = byte_range_to_char_range(line, value.start(), value.end());
-            blank_range(&mut chars, start, end);
-        }
-    }
-
-    mask_inline_code(&line_chars, &mut chars);
-    mask_link_targets(&line_chars, &mut chars);
-
-    for value in &mut chars {
-        if matches!(*value, '\\' | '*' | '_' | '~' | '|' | '!' | '[' | ']' | '(' | ')') {
-            *value = ' ';
-        }
-    }
-
-    chars.into_iter().collect()
-}
-
-fn mask_inline_code(line_chars: &[char], chars: &mut [char]) {
-    let mut index = 0;
-
-    while index < line_chars.len() {
-        if line_chars[index] != '`' {
-            index += 1;
-            continue;
-        }
-
-        let mut tick_count = 1;
-        while index + tick_count < line_chars.len() && line_chars[index + tick_count] == '`' {
-            tick_count += 1;
-        }
-
-        if let Some(close_index) = find_backtick_fence(line_chars, index + tick_count, tick_count) {
-            blank_range(chars, index, close_index + tick_count);
-            index = close_index + tick_count;
-        } else {
-            blank_range(chars, index, index + tick_count);
-            index += tick_count;
-        }
-    }
-}
-
-fn mask_link_targets(line_chars: &[char], chars: &mut [char]) {
-    let mut index = 0;
-
-    while index + 1 < line_chars.len() {
-        if line_chars[index] != ']' {
-            index += 1;
-            continue;
-        }
-
-        if line_chars[index + 1] == '(' {
-            let mut depth = 1_i32;
-            let mut cursor = index + 2;
-            while cursor < line_chars.len() && depth > 0 {
-                if line_chars[cursor] == '(' {
-                    depth += 1;
-                } else if line_chars[cursor] == ')' {
-                    depth -= 1;
-                }
-                cursor += 1;
-            }
-            blank_range(chars, index, cursor);
-            index = cursor;
-            continue;
-        }
-
-        if line_chars[index + 1] == '[' {
-            let mut cursor = index + 2;
-            while cursor < line_chars.len() {
-                if line_chars[cursor] == ']' {
-                    blank_range(chars, index, cursor + 1);
-                    index = cursor + 1;
-                    break;
-                }
-                cursor += 1;
-            }
-        } else {
-            index += 1;
-        }
-    }
-}
-
-fn find_backtick_fence(chars: &[char], start: usize, tick_count: usize) -> Option<usize> {
-    let mut index = start;
-
-    while index + tick_count <= chars.len() {
-        if chars[index..index + tick_count].iter().all(|value| *value == '`') {
-            return Some(index);
-        }
-        index += 1;
-    }
-
-    None
-}
-
-fn blank_range(chars: &mut [char], start: usize, end: usize) {
-    let safe_start = start.min(chars.len());
-    let safe_end = end.min(chars.len());
-    for value in chars.iter_mut().take(safe_end).skip(safe_start) {
-        *value = ' ';
-    }
-}
-
-fn byte_range_to_char_range(text: &str, start: usize, end: usize) -> (usize, usize) {
-    (byte_to_char_index(text, start), byte_to_char_index(text, end))
-}
-
-fn byte_to_char_index(text: &str, byte_index: usize) -> usize {
-    text[..byte_index.min(text.len())].chars().count()
-}
-
-fn collect_char_boundaries(text: &str) -> Vec<usize> {
-    let mut boundaries = text.char_indices().map(|(index, _)| index).collect::<Vec<_>>();
-    boundaries.push(text.len());
-    boundaries
-}
-
-fn sort_and_dedupe_segment_words(words: &mut Vec<SegmentWord>) {
-    words.sort_by(|left, right| {
-        right.char_len.cmp(&left.char_len).then_with(|| left.text.cmp(&right.text))
-    });
-    words.dedup_by(|left, right| left.text == right.text);
-}
-
-fn normalize_comparable_word(word: &str) -> String {
-    normalize_latin_word(word).chars().filter(|value| !matches!(value, '\'' | '’' | '-')).collect()
-}
-
-fn normalize_word_for_set(word: &str) -> String {
-    if word.chars().any(is_cjk_char) {
-        word.nfc().collect::<String>().trim().to_string()
-    } else {
-        normalize_latin_word(word)
-    }
-}
-
-fn normalize_latin_word(word: &str) -> String {
-    word.nfc().flat_map(char::to_lowercase).collect()
-}
-
-fn collapse_whitespace(text: &str) -> String {
-    let mut collapsed = CompactString::default();
-    let mut needs_space = false;
-
-    for part in text.split_whitespace() {
-        if needs_space {
-            collapsed.push(' ');
-        }
-        collapsed.push_str(part);
-        needs_space = true;
-    }
-
-    collapsed.into_string()
-}
-
-fn count_code_points(text: &str) -> usize {
-    text.chars().count()
-}
-
-fn dedupe_strings(values: Vec<String>) -> Vec<String> {
-    let mut seen = FxHashSet::default();
-    let mut deduped = Vec::new();
-
-    for value in values {
-        if seen.insert(value.clone()) {
-            deduped.push(value);
-        }
-    }
-
-    deduped
-}
-
-fn is_supported_language(language: &str) -> bool {
-    SUPPORTED_MARKDOWN_LINT_LANGUAGES.contains(&language)
-}
-
-fn is_repeated_punctuation_char(value: char) -> bool {
-    matches!(value, '!' | '?' | '！' | '？' | '。' | '、' | '，')
-}
-
-fn is_uppercase_token(value: &str) -> bool {
-    let mut chars = value.chars();
-    let Some(first_char) = chars.next() else {
-        return false;
-    };
-
-    first_char.is_uppercase()
-        && chars.all(|character| character.is_uppercase() || character.is_numeric())
-}
-
-fn is_hiragana(value: char) -> bool {
-    matches!(value, '\u{3040}'..='\u{309F}')
-}
-
-fn is_cjk_char(value: char) -> bool {
-    matches!(
-        value,
-        '\u{3040}'..='\u{309F}' | '\u{30A0}'..='\u{30FF}' | '\u{4E00}'..='\u{9FFF}'
-    )
-}
-
-fn levenshtein(left: &str, right: &str) -> usize {
-    if left == right {
-        return 0;
-    }
-
-    let left_chars = left.chars().collect::<Vec<_>>();
-    let right_chars = right.chars().collect::<Vec<_>>();
-
-    if left_chars.is_empty() {
-        return right_chars.len();
-    }
-    if right_chars.is_empty() {
-        return left_chars.len();
-    }
-
-    let mut previous_row = (0..=right_chars.len()).collect::<Vec<_>>();
-    let mut current_row = vec![0; right_chars.len() + 1];
-
-    for (left_index, left_value) in left_chars.iter().enumerate() {
-        current_row[0] = left_index + 1;
-
-        for (right_index, right_value) in right_chars.iter().enumerate() {
-            let substitution_cost = usize::from(left_value != right_value);
-            let insertion = current_row[right_index] + 1;
-            let deletion = previous_row[right_index + 1] + 1;
-            let substitution = previous_row[right_index] + substitution_cost;
-            current_row[right_index + 1] = insertion.min(deletion).min(substitution);
-        }
-
-        previous_row.clone_from(&current_row);
-    }
-
-    previous_row[right_chars.len()]
 }
